Create MockObjects file for Common tests
